Updated Movement, Perspective Camera, & Smaller Maze
Changed the player camera back to perspective, orthographic was clipping lights. The maze is a little smaller and the level now only has a dim directional light to help with guaging the player's point light that will be implemented. There is a nav mesh but right now the player is just being prevented from walking off edges